+++ jnalib/native/libffi/testsuite/libffi.call/cls_pointer_stack.c
@@ -20,43 +20,48 @@ long dummy_func(
 
 void* cls_pointer_fn2(void* a1, void* a2)
 {
-	long double	trample1	= (long)a1 + (long)a2;
+	long double	trample1	= (intptr_t)a1 + (intptr_t)a2;
 	char		trample2	= ((char*)&a1)[0] + ((char*)&a2)[0];
-	long double	trample3	= (long)trample1 + (long)a1;
+	long double	trample3	= (intptr_t)trample1 + (intptr_t)a1;
 	char		trample4	= trample2 + ((char*)&a1)[1];
-	long double	trample5	= (long)trample3 + (long)a2;
+	long double	trample5	= (intptr_t)trample3 + (intptr_t)a2;
 	char		trample6	= trample4 + ((char*)&a2)[1];
-	long double	trample7	= (long)trample5 + (long)trample1;
+	long double	trample7	= (intptr_t)trample5 + (intptr_t)trample1;
 	char		trample8	= trample6 + trample2;
 
 	dummyVar	= dummy_func(trample1, trample2, trample3, trample4,
 		trample5, trample6, trample7, trample8);
 
-	void*	result	= (void*)((long)a1 + (long)a2);
+	void*	result	= (void*)((intptr_t)a1 + (intptr_t)a2);
 
 	printf("0x%08x 0x%08x: 0x%08x\n", 
-	       (unsigned int) a1, (unsigned int) a2, (unsigned int) result);
+	       (unsigned int)(uintptr_t) a1,
+               (unsigned int)(uintptr_t) a2,
+               (unsigned int)(uintptr_t) result);
 
 	return result;
 }
 
 void* cls_pointer_fn1(void* a1, void* a2)
 {
-	long double	trample1	= (long)a1 + (long)a2;
+	long double	trample1	= (intptr_t)a1 + (intptr_t)a2;
 	char		trample2	= ((char*)&a1)[0] + ((char*)&a2)[0];
-	long double	trample3	= (long)trample1 + (long)a1;
+	long double	trample3	= (intptr_t)trample1 + (intptr_t)a1;
 	char		trample4	= trample2 + ((char*)&a1)[1];
-	long double	trample5	= (long)trample3 + (long)a2;
+	long double	trample5	= (intptr_t)trample3 + (intptr_t)a2;
 	char		trample6	= trample4 + ((char*)&a2)[1];
-	long double	trample7	= (long)trample5 + (long)trample1;
+	long double	trample7	= (intptr_t)trample5 + (intptr_t)trample1;
 	char		trample8	= trample6 + trample2;
 
 	dummyVar	= dummy_func(trample1, trample2, trample3, trample4,
 		trample5, trample6, trample7, trample8);
 
-	void*	result	= (void*)((long)a1 + (long)a2);
+	void*	result	= (void*)((intptr_t)a1 + (intptr_t)a2);
 
-	printf("0x%08x 0x%08x: 0x%08x\n", a1, a2, result);
+	printf("0x%08x 0x%08x: 0x%08x\n",
+               (unsigned int)(intptr_t) a1,
+               (unsigned int)(intptr_t) a2,
+               (unsigned int)(intptr_t) result);
 
 	result	= cls_pointer_fn2(result, a1);
 
@@ -70,13 +75,13 @@ cls_pointer_gn(ffi_cif* cif __UNUSED__, void* resp,
 	void*	a1	= *(void**)(args[0]);
 	void*	a2	= *(void**)(args[1]);
 
-	long double	trample1	= (long)a1 + (long)a2;
+	long double	trample1	= (intptr_t)a1 + (intptr_t)a2;
 	char		trample2	= ((char*)&a1)[0] + ((char*)&a2)[0];
-	long double	trample3	= (long)trample1 + (long)a1;
+	long double	trample3	= (intptr_t)trample1 + (intptr_t)a1;
 	char		trample4	= trample2 + ((char*)&a1)[1];
-	long double	trample5	= (long)trample3 + (long)a2;
+	long double	trample5	= (intptr_t)trample3 + (intptr_t)a2;
 	char		trample6	= trample4 + ((char*)&a2)[1];
-	long double	trample7	= (long)trample5 + (long)trample1;
+	long double	trample7	= (intptr_t)trample5 + (intptr_t)trample1;
 	char		trample8	= trample6 + trample2;
 
 	dummyVar	= dummy_func(trample1, trample2, trample3, trample4,
